+++ packages/gallery/src/components/layoutFixer/layoutFixer.js
@@ -12,11 +12,11 @@ export const LayoutFixer = (props) => {
             <style id={'layout-fixer-style-' + props.domId}></style>
             {isPrerenderMode() ? (
                 <>
-                    {(props.layoutFixerScriptType !== 'nopreload') && <link rel="preload" href={props.layoutFixerBundleUrl} as="script"/>}
+                    {(props.layoutFixerScriptType === 'preload') && <link rel="preload" href={props.layoutFixerBundleUrl} as="script"/>}
                     <script dangerouslySetInnerHTML={{__html: `console.log("[LAYOUT FIXER] Start (first script loaded)")`}}></script>
-                    {props.layoutFixerScriptType === 'async' && <script async type="module" src={props.layoutFixerBundleUrl}></script>}
+                    {props.layoutFixerScriptType === 'async' || props.layoutFixerScriptType === 'preload' && <script async type="module" src={props.layoutFixerBundleUrl}></script>}
                     {props.layoutFixerScriptType === 'defer' && <script defer src={props.layoutFixerBundleUrl}></script>}
-                    {(props.layoutFixerScriptType !== 'async' && props.layoutFixerScriptType !== 'defer') && <script src={props.layoutFixerBundleUrl}></script>}
+                    {(props.layoutFixerScriptType !== 'async' && props.layoutFixerScriptType !== 'defer' && props.layoutFixerScriptType !== 'preload') && <script src={props.layoutFixerBundleUrl}></script>}
                     <layout-fixer 
                         domId={props.domId}
                         items={JSON.stringify(props.items)}