+++ packages/gallery/tests/e2e/layouts/empty.e2e.spec.js
@@ -7,16 +7,16 @@ expect.extend({ toMatchImageSnapshot });
 describe('empty - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('empty - scrollDirection = vertical', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL
     });
@@ -26,7 +26,7 @@ describe('empty - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('empty - scrollDirection = horizontal', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       scrollDirection: GALLERY_CONSTS.scrollDirection.HORIZONTAL
     });