+++ packages/gallery/tests/e2e/styleParams/fixedColumns.e2e.spec.js
@@ -7,42 +7,45 @@ expect.extend({ toMatchImageSnapshot });
 describe('fixedColumns - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('should set fixedColimns of "1"', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       groupSize:1,
       isVertical:true,
       fixedColumns: 1,
+      gridStyle: 1,
     });
     await driver.waitFor.hookToBeVisible('item-container');
     const page = await driver.grab.partialScreenshot();
     expect(page).toMatchImageSnapshot();
   });
   it('should set fixedColimns of "2"', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       groupSize:1,
       isVertical:true,
       fixedColumns: 2,
+      gridStyle: 1,
     });
     await driver.waitFor.hookToBeVisible('item-container');
     const page = await driver.grab.elemScreenshot('.pro-gallery');
     expect(page).toMatchImageSnapshot();
   });
   it('should set fixedColimns of "3"', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       groupSize:1,
       isVertical:true,
       fixedColumns: 3,
+      gridStyle: 1,
     });
     await driver.waitFor.hookToBeVisible('item-container');
     const page = await driver.grab.elemScreenshot('.pro-gallery');