+++ packages/gallery/tests/e2e/styleParams/groupSize.e2e.spec.js
@@ -7,16 +7,16 @@ expect.extend({ toMatchImageSnapshot });
 describe('groupSize - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('should have max group size of 3 (groupSize=3)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       collageDensity: 0.8,
       groupSize: 3,
@@ -27,7 +27,7 @@ describe('groupSize - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should have max group size of 1 (groupSize=1)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       collageDensity: 0.8,
       groupSize: 1,
@@ -38,7 +38,7 @@ describe('groupSize - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should have groups of 1 item (restricted by collageDensity)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       collageDensity: 0,
       groupSize: 3,
@@ -49,9 +49,10 @@ describe('groupSize - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should have groups of 1 item (restricted by groupsPerStrip)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       groupsPerStrip: 1,
+      gridStyle: 1,
       groupSize: 3,
     });
     await driver.waitFor.hookToBeVisible('item-container');