+++ packages/gallery/tests/e2e/styleParams/hasThumbnails.e2e.spec.js
@@ -7,16 +7,16 @@ expect.extend({ toMatchImageSnapshot });
 describe('hasThumbnails - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('should not have thumbnails when "hasThumbnails" is "false" and gallery is horizontal scroll', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       hasThumbnails: false,
       oneRow:true,
@@ -27,7 +27,7 @@ describe('hasThumbnails - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should not have thumbnails when "hasThumbnails" is "true" and gallery is vertical scroll', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       hasThumbnails: false,
     });
@@ -37,7 +37,7 @@ describe('hasThumbnails - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should have thumbnails when "hasThumbnails" is "true" and gallery is horizontal scroll', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.EMPTY,
       hasThumbnails: true,
       oneRow:true,