+++ packages/gallery/tests/e2e/styleParams/imageInfoType.e2e.spec.js
@@ -7,16 +7,16 @@ expect.extend({ toMatchImageSnapshot });
 describe('imageInfoType - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('should apply styles to image only (imageInfoType = "NO_BACKGROUND")', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       imageInfoType: GALLERY_CONSTS.infoType.NO_BACKGROUND,
       itemBorderWidth: 5,
@@ -31,7 +31,7 @@ describe('imageInfoType - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should attach texts and image and apply styles to both as one container (imageInfoType = "ATTACHED_BACKGROUND")', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       imageInfoType: GALLERY_CONSTS.infoType.ATTACHED_BACKGROUND,
       itemBorderWidth: 5,
@@ -46,7 +46,7 @@ describe('imageInfoType - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should separate texts and image (imageInfoType = "SEPARATED_BACKGROUND")', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       imageInfoType: GALLERY_CONSTS.infoType.SEPARATED_BACKGROUND,
       itemBorderWidth: 5,
@@ -61,7 +61,7 @@ describe('imageInfoType - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should not show styles to texts and image (imageInfoType = "SEPARATED_BACKGROUND")', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       imageInfoType: GALLERY_CONSTS.infoType.DONT_SHOW,
       itemBorderWidth: 5,