+++ packages/gallery/tests/e2e/styleParams/textImageSpace.e2e.spec.js
@@ -7,16 +7,16 @@ expect.extend({ toMatchImageSnapshot });
 describe('textImageSpace - e2e', () => {
   let driver;
 
-  beforeEach(async () => {
+  beforeAll(async () => {
     driver = new GalleryDriver();
-    await driver.launchBrowser();
+    await driver.openPage();
   });
 
-  afterEach(() => {
-    driver.closeBrowser();
+  afterAll( async() => {
+    await driver.closePage();
   });
   it('should set spacing between texts and image (texts below)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
       imageInfoType: GALLERY_CONSTS.infoType.SEPARATED_BACKGROUND,
@@ -28,7 +28,7 @@ describe('textImageSpace - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should set spacing between texts and image (texts above)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_ABOVE,
       imageInfoType: GALLERY_CONSTS.infoType.SEPARATED_BACKGROUND,
@@ -40,7 +40,7 @@ describe('textImageSpace - e2e', () => {
     expect(page).toMatchImageSnapshot();
   });
   it('should not set spacing between texts and image (texts above, no separated background)', async () => {
-    await driver.openPage({
+    await driver.navigate({
       galleryLayout: GALLERY_CONSTS.layout.GRID,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
       imageInfoType: GALLERY_CONSTS.infoType.NO_BACKGROUND,