+++ packages/lib/src/settings/options/gridStyle.js
@@ -4,11 +4,11 @@ import { createOptions } from '../utils/utils';
 
 export default {
   title: 'Responsive Type',
-  isRelevant: (styleParams) => styleParams.isVertical,
+  isRelevant: (styleParams) => !styleParams.oneRow,
   type: INPUT_TYPES.OPTIONS,
   default: GALLERY_CONSTS.gridStyle.FIT_TO_SCREEN,
   options: createOptions('gridStyle'),
   description: `Choose between adjusting the number of columns addording to the container
-  size or setting it manualyand keep it fixed.
-  `,
+  size or setting it manually and keep it fixed.`,
+  alert: 'This option will disable the responsive feature of the gallery!',
 };