+++ packages/lib/src/settings/options/rotatingCropRatios.js
@@ -1,7 +1,7 @@
 import { INPUT_TYPES } from '../utils/constants';
 
 export default {
-  title: 'Rotating Crop Ratios',
+  title: 'Fixed Crop Ratios',
   description: `Crop each image according to the corresponding crop ratio as indicated in this string. This will create a pattern of cropped images`,
   isRelevant: (styleParams) => styleParams.cubeImages,
   type: INPUT_TYPES.TEXT,