+++ packages/lib/src/settings/options/scrollAnimation.js
@@ -5,7 +5,7 @@ import { createOptions } from '../utils/utils';
 export default {
   title: 'Scroll Animation',
   description: `Choose the type of animation to be used when items appear while scrolling verticaly through the gallery`,
-  isRelevant: () => true,
+  isRelevant: (sp) => !sp.oneRow || sp.slideAnimation === GALLERY_CONSTS.slideAnimations.SCROLL,
   options: createOptions('scrollAnimations'),
   type: INPUT_TYPES.OPTIONS,
   default: GALLERY_CONSTS.scrollAnimations.NO_EFFECT,