+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## [v2.2.19](https://pro-gallery-2-2-19.surge.sh) (11/11/2020)
+ 
+#### GALLERY
+ -  fix (videoItem): add namespace webpack chunk names (to avoid react-player imports collisions)
+ 
+---
+## [v2.2.18](https://pro-gallery-2-2-18.surge.sh) (11/11/2020)
+ 
+#### LAYOUTS
+ -  fix: keep old hashToInt logic for grouping
+
+#### MAIN
+ -  moved CI to travis-ci.com
+
+---
+## [v2.2.17](https://pro-gallery-2-2-17.surge.sh) (11/11/2020)
+ 
+#### GALLERY
+ -  fix (slideshowview): load non-visible items before thumbnail navigation
+ -  fix (itemView): remove double borderRadius for attached background
+ -  improve (scrollAnimations): enable and fix scroll animations in horizontal galleries
+
+#### LAYOUTS
+ -  improve: better scatter calculation
+
+---
+## [v2.2.16](https://pro-gallery-2-2-16.surge.sh) (09/11/2020)
+ 
+#### GALLERY
+ -  improve (layoutFixer): remove default preload option
+
+---
+## [v2.2.15](https://pro-gallery-2-2-15.surge.sh) (08/11/2020)
+ 
+#### GALLERY
+ -  Fixed condition to preview the hover element.
+
+#### PLAYGROUND
+ -  improve: add scroll animations for horizontal galleries
+
+---
 ## [v2.2.14](https://pro-gallery-2-2-14.surge.sh) (05/11/2020)
  
 #### GALLERY